Move flyback calculations to topologies
